# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: IntelliDesk API
  version: 0.1.0
  description: |
    与 PRD_Backend_Detailed 对齐。鉴权：请求头 `X-Tenant-Id`（必填）、`X-User-Id`（可选，演示用已存在用户 id）。
servers:
  - url: http://localhost:4000
    description: 本地后端直连（默认 PORT）
  - url: http://localhost:4001
    description: 经前端 Vite 同源代理（npm run dev:api，浏览器站点根）
...
components:
  securitySchemes:
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
  parameters:
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
security:
  - TenantHeader: []
paths:
...
  /api/orders/{id}:
    get:
      tags: [orders]
      summary: 订单详情
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: upstreamEnrich
          in: query
          description: 设为 1 时同步请求母系统补全地址/单号（默认不请求，缩短响应）
          schema: { type: string, enum: ['0', '1'] }
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Order DTO }
        '404': { description: Not found }
  /api/orders/{id}/logistics:
    get:
      tags: [orders]
      summary: 物流事件
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: events array }
...